58 outdated entities

+++ composeApp/src/commonMain/kotlin/plus/vplan/app/data/repository/AssessmentRepositoryImpl.kt
@@ -29,6 +29,7 @@ import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.flow.takeWhile
 import kotlinx.coroutines.launch
+import kotlinx.datetime.Clock
 import kotlinx.datetime.Instant
 import kotlinx.datetime.LocalDate
 import kotlinx.datetime.TimeZone
@@ -39,6 +40,7 @@ import plus.vplan.app.api
 import plus.vplan.app.data.source.database.VppDatabase
 import plus.vplan.app.data.source.database.model.database.DbAssessment
 import plus.vplan.app.data.source.database.model.database.foreign_key.FKAssessmentFile
+import plus.vplan.app.data.source.network.isResponseFromBackend
 import plus.vplan.app.data.source.network.safeRequest
 import plus.vplan.app.data.source.network.toErrorResponse
 import plus.vplan.app.data.source.network.toResponse
@@ -91,6 +93,7 @@ class AssessmentRepositoryImpl(
                     defaultLessonId = assessment.subject,
                     description = assessment.description,
                     type = (Assessment.Type.entries.firstOrNull { it.name == assessment.type } ?: Assessment.Type.OTHER).ordinal,
+                    cachedAt = Clock.System.now()
                 ) },
                 files = assessments.flatMap { assessment ->
                     assessment.files.map {
@@ -175,6 +178,10 @@ class AssessmentRepositoryImpl(
         return vppDatabase.assessmentDao.getAll().map { it.map { item -> item.toModel() } }
     }
 
+    override fun getAllIds(): Flow<List<Int>> {
+        return vppDatabase.assessmentDao.getAll().map { it.map { it.assessment.id } }
+    }
+
     override fun getById(id: Int, forceReload: Boolean): Flow<CacheState<Assessment>> {
         if (id < 0) {
             return vppDatabase.assessmentDao.getById(id).map {
@@ -199,7 +206,7 @@ class AssessmentRepositoryImpl(
             ) {
                 trySend(CacheState.Loading(id.toString())).onFailure { return@channelFlow }
                 val metadataResponse = httpClient.get("${api.url}/api/v2.2/assessment/$id")
-                if (metadataResponse.status == HttpStatusCode.NotFound) {
+                if (metadataResponse.status == HttpStatusCode.NotFound && metadataResponse.isResponseFromBackend()) {
                     trySend(CacheState.NotExisting(id.toString()))
                     vppDatabase.assessmentDao.deleteById(listOf(id))
                     return@channelFlow
@@ -247,7 +254,8 @@ class AssessmentRepositoryImpl(
                         isPublic = data.isPublic,
                         defaultLessonId = data.subject,
                         description = data.description,
-                        type = (Assessment.Type.entries.firstOrNull { it.name == data.type } ?: Assessment.Type.OTHER).ordinal
+                        type = (Assessment.Type.entries.firstOrNull { it.name == data.type } ?: Assessment.Type.OTHER).ordinal,
+                        cachedAt = Clock.System.now()
                     )),
                     files = data.files.map { FKAssessmentFile(id, it) }
                 )
@@ -296,7 +304,8 @@ class AssessmentRepositoryImpl(
                 isPublic = it.isPublic,
                 defaultLessonId = it.defaultLessonId,
                 description = it.description,
-                type = it.type.ordinal
+                type = it.type.ordinal,
+                cachedAt = it.cachedAt
             ) },
             files = assessments.flatMap { assessment ->
                 assessment.files.map { fileId -> FKAssessmentFile(
